nzcarwebapp/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
setuptools==65.5.0
Werkzeug==2.3.6
